Fix timeout exception on Network create
In fact `requests` `connect` timeout is *not* just TCP connect, but
also includes sending the whole HTTP request, which can be much
longer than 3s for Network uploads.

Before the issue is fixed in `requests` we should not fail long
uploads: stop using advanced timeouts for now.

Also decreased FAST timeout from 10 to 8 so we are guaranteed to have
at least 3 tries with the 60s timeout on tasks GET loop.